refactor: change the structure of the project (author and main page)
