+++ src/components/authorPage/YandexMap/yandex-map-style.js
