+++ src/components/authorPage/artist's-works/StyledTableCell/StyledTableCell.js
