+++ src/components/authorPage/artist's-works/artistSWorksStyles.jsx
