+++ src/components/authorPage/image-gallery/PhotoGallery.jsx
