+++ src/components/authorPage/image-gallery/image-gallery.css
