+++ src/components/authorPage/video/videoStyles.jsx
